+++ src/components/TopBar/TopBar.tsx
@@ -36,6 +36,7 @@ import Avatar from '@material-ui/core/Avatar'
 class TopBar extends React.Component <{
 	showAlert: (message: string, severity: 'success' | 'warning' | 'info' | 'error' | undefined) => void
 	onSelectedCourse: (selectedCourses: Course[]) => Promise<void>
+	onSelectedDegree: (selectedDegrees: Degree[]) => Promise<void>
 	onClearShifts: (alert: boolean) => void
 	onGetLink: () => void
 	onChangeLanguage: (language: string) => void
@@ -53,7 +54,9 @@ class TopBar extends React.Component <{
 		warningDialog: false,
 		languageAnchor: null
 	}
-	selectedDegree: Degree | null = null
+	selectedDegrees: Degree[] = []
+	tempSelectedDegrees: string[] = []
+	hasPreviousDegrees = false
 
 	// eslint-disable-next-line
 	constructor(props: any) {
@@ -72,18 +75,33 @@ class TopBar extends React.Component <{
 		if (degrees === null) {
 			this.props.showAlert(i18next.t('alert.cannot-obtain-degrees'), 'error')
 		}
+		if (this.tempSelectedDegrees.length > 0) {
+			this.setSelectedDegrees(this.tempSelectedDegrees)
+			this.tempSelectedDegrees = []
+		}
 	}
 
-	async onSelectedDegree(degree: Degree | null): Promise<void> {
-		this.selectedDegree = degree
-		if (degree !== null) {
-			const degreeCourses = await API.getCourses(degree) 
-			if (degreeCourses === null) {
-				this.props.showAlert(i18next.t('alert.cannot-obtain-courses'), 'error')
-				return
+	async onSelectedDegree(degrees: Degree[]): Promise<void> {
+		this.selectedDegrees = degrees
+		if (degrees.length > 0) {
+			let degreeCourses: Course[] = []
+			for (const degree of degrees) {
+				const tempCourses = await API.getCourses(degree) 
+				if (tempCourses === null) {
+					// TODO: Test when this cannot be obtained
+					this.props.showAlert(i18next.t('alert.cannot-obtain-courses'), 'error')
+					return
+				}
+				degreeCourses = degreeCourses.concat(tempCourses)
 			}
 			const selected = this.state.selectedCourses.courses
 			const availableCourses = Comparables.toUnique(degreeCourses.concat(selected)) as Course[]
+
+			if (this.hasPreviousDegrees) {
+				// If degrees were loaded, update courses to have the right degree
+				availableCourses.forEach((c) => c.updateDegree(this.selectedDegrees.map(d => d.acronym)))
+				this.hasPreviousDegrees = false
+			}
 			this.setState({
 				availableCourses: availableCourses.sort(Course.compare)
 			})
@@ -92,6 +110,7 @@ class TopBar extends React.Component <{
 				availableCourses: this.state.selectedCourses.courses
 			})
 		}
+		this.props.onSelectedDegree(this.selectedDegrees)
 	}
 
 	setHasSelectedShifts(shifts: Shift[]): void {
@@ -103,6 +122,17 @@ class TopBar extends React.Component <{
 	//FIXME: Available courses not updating when a course from another degree is removed 
 	private async onSelectedCourse(selectedCourses: Course[]): Promise<void> {
 		this.props.onSelectedCourse(selectedCourses)
+	}
+
+	setSelectedDegrees(selectedDegreesAcronyms: string[]): void {
+		// Store temporarily the degrees for the degrees to be loaded
+		if (this.state.degrees.length === 0) {
+			this.hasPreviousDegrees = true
+			this.tempSelectedDegrees = selectedDegreesAcronyms
+		} else { // When degrees are loaded, select them
+			this.selectedDegrees = this.state.degrees.filter( (d) => selectedDegreesAcronyms.includes(d.acronym))
+			this.onSelectedDegree(this.selectedDegrees)
+		}
 	}
 
 	setSelectedCourses(selectedCourses: CourseUpdates): void {
@@ -124,7 +154,7 @@ class TopBar extends React.Component <{
 		}
 
 		this.onSelectedCourse([])
-		this.onSelectedDegree(this.selectedDegree)
+		this.onSelectedDegree(this.selectedDegrees)
 		this.props.onClearShifts(false)
 		this.setState({
 			selectedAcademicTerm: s
@@ -153,7 +183,7 @@ class TopBar extends React.Component <{
 			degrees: degrees ?? []
 		})
 
-		this.selectedDegree = null
+		this.selectedDegrees = []
 		this.onSelectedCourse([])
 	}
 
@@ -176,10 +206,11 @@ class TopBar extends React.Component <{
 				>
 					<Toolbar className={styles.ToolBar}>
 						<Autocomplete
-							value={this.selectedDegree}
+							value={this.selectedDegrees}
 							color="inherit"
 							size="small"
 							className={styles.degreeSelector}
+							multiple
 							selectOnFocus
 							clearOnBlur
 							handleHomeEndKeys={false}
@@ -188,6 +219,14 @@ class TopBar extends React.Component <{
 							options={this.state.degrees}
 							getOptionLabel={(option) => option.displayName()}
 							renderInput={(params) => <TextField {...params} label={i18next.t('degree-selector.title') as string} variant="outlined" />}
+							renderTags={(tagValue, getTagProps) => {
+								// TODO: Fix chip color
+								return tagValue.map((option, index) => (
+									<Tooltip title={option.displayName()} key={option.hashString()}>
+										<Chip {...getTagProps({ index })} size="small" color='secondary' label={option.acronym} />
+									</Tooltip>
+								))
+							}}
 						/>
 						<Autocomplete
 							value={this.state.selectedCourses.courses}
@@ -207,7 +246,11 @@ class TopBar extends React.Component <{
 									this.state.selectedCourses.courses.length === maxSelectedCourses
 							}}
 							noOptionsText={i18next.t('course-selector.noOptions') as string}
-							getOptionLabel={(option) => option.displayName()}
+							getOptionLabel={(option) => {
+								// Make course show degree when multiple are chosen
+								option.showDegree = this.selectedDegrees.length > 1
+								return option.displayName()
+							}}
 							renderInput={(params) => <TextField  {...params} label={i18next.t('course-selector.title') as string} variant="outlined" />}
 							renderTags={(tagValue, getTagProps) => {
 								return tagValue.map((option, index) => (