+++ src/domain/Course.tsx
@@ -103,10 +103,11 @@ export default class Course implements Comparable {
 	}
 
 	searchableName(): string {
-		return this.abbrev + this.name + this.acronym
+		return this.abbrev + this.name + this.acronym +
+			(this.showDegree ? this.degreeAcronym : '') // Add degree to be able to only show by degree
 	}
 
-	displayName() : string {
+	displayName(): string {
 		let display = this.name
 		if (this.showDegree) {
 			if (this.degreeAcronym !== '') {
@@ -116,6 +117,21 @@ export default class Course implements Comparable {
 			}
 		}
 		return display
+	}
+
+	updateDegree(degreesAcronyms: string[]): void {
+		if (this.degreeAcronym.length === 0 || degreesAcronyms.length === 0) {
+			// How did this happen???
+		} else {
+			if (degreesAcronyms.length > 1) {
+				this.showDegree = true
+			}
+			// Filter out unselected degrees when coming from previous sessions
+			const possibleDegrees = this.degreeAcronym.split('/').filter((a) => degreesAcronyms.includes(a))
+			if (possibleDegrees.length > 0) { // Only change if has a possible degree
+				this.degreeAcronym = possibleDegrees.join('/')
+			}
+		}
 	}
 }