+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 28/06/2021
+
+### Added
+
+- Allow selection of multiple degrees to have all the courses in the selection list
+- Allow storing degrees in cookies and urls for sharing and showing proper classes
+- Add minimal classes functionality when getting the classes
+
+### Fixed
+
+- Add temporary favicon.ico
+- When unselecting a course that has one or more shifts selected and then selecting again, the chips do not mark as selected the corresponding shifts
+
 ## [1.5.0] - 02/03/2021
 
 ### Added